chore: Resume/Portfolio의 디렉토리 구조 변경
- Document 디렉토리 하위에 있던 Resume과 Portfolio를 각각 Document와 같은 level로 resume, portfolio 디렉토리를 만들어 옮겼습니다.

- Document의 하위 도메인은 맞으나, coverletter와 동일하게 위치시키는 게 맞는 것 같아 늦게라도 이동합니다.
