+++ src/main/java/com/careerfit/portfolio/repository/PortfolioRepository.java
@@ -1,6 +1,6 @@
-package com.careerfit.document.repository;
+package com.careerfit.portfolio.repository;
 
-import com.careerfit.document.domain.Portfolio;
+import com.careerfit.portfolio.domain.Portfolio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
