+++ src/main/java/com/careerfit/resume/repository/ResumeRepository.java
@@ -1,6 +1,6 @@
-package com.careerfit.document.repository;
+package com.careerfit.resume.repository;
 
-import com.careerfit.document.domain.Resume;
+import com.careerfit.resume.domain.Resume;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ResumeRepository extends JpaRepository<Resume, Long> {
